+++ src/interfaces/rate-limiter-options.ts
@@ -1,3 +1,12 @@
+interface Rate {
+  max: number;
+  duration: number;
+}
+
+export interface GroupRates {
+  [key: string]: Rate;
+}
+
 export interface RateLimiterOptions {
   // Max number of jobs processed
   max: number;
@@ -7,4 +16,7 @@ export interface RateLimiterOptions {
 
   // grouping path key in job data
   groupKey?: string;
+
+  // optional rate limits defined per groupKey
+  groupRates?: GroupRates;
 }