+++ src/test/test_rate_limiter.ts
@@ -8,6 +8,7 @@ import { after, every, last } from 'lodash';
 import { beforeEach, describe, it } from 'mocha';
 import { v4 } from 'uuid';
 import { removeAllQueueData } from '@src/utils';
+import { RateLimiterOptions } from '@src/interfaces';
 
 describe('Rate Limiter', function() {
   let queue: Queue;
@@ -116,18 +117,18 @@ describe('Rate Limiter', function() {
     const queueScheduler = new QueueScheduler(queueName);
     await queueScheduler.waitUntilReady();
 
+    const limiterConfig: RateLimiterOptions = {
+      max: 1,
+      duration: 1000,
+      groupKey: 'accountId',
+    };
+
     const rateLimitedQueue = new Queue(queueName, {
-      limiter: {
-        groupKey: 'accountId',
-      },
+      limiter: limiterConfig,
     });
 
     const worker = new Worker(queueName, async job => {}, {
-      limiter: {
-        max: 1,
-        duration: 1000,
-        groupKey: 'accountId',
-      },
+      limiter: limiterConfig,
     });
 
     const completed: { [index: string]: number[] } = {};
@@ -173,6 +174,120 @@ describe('Rate Limiter', function() {
     }
 
     await running;
+    await rateLimitedQueue.close();
+    await worker.close();
+    await queueScheduler.close();
+  });
+
+  it('should respect separate rate limits by grouping', async function() {
+    this.timeout(20000);
+
+    const numGroups = 3;
+    const numJobsPerGroup = 5;
+    const startTime = Date.now();
+    const buffer = 10; // Used for time diffs, sometimes things are off by a few ms
+
+    const queueScheduler = new QueueScheduler(queueName);
+    await queueScheduler.waitUntilReady();
+
+    // NOTE:
+    // The groupRates feature is a little weird because Queue and Worker require the same
+    // rate limiter options for it to work.
+    // Its because Queue is used for addJob, which adds the group-specific rate limits
+    // for whatever job we're adding.
+    // However, Worker is used for moveToActive, which technically doesn't need the
+    // separate group rate limits defined because those will have already been in Redis,
+    // but it does need groupKey defined to enable group rate limiting logic, and also
+    // the default max/duration so it can fallback to that if group rate limits aren't
+    // found.
+    const limiterConfig: RateLimiterOptions = {
+      max: 1,
+      duration: 3000, // This default should apply for 'group3' jobs
+      groupKey: 'accountId',
+      groupRates: {
+        group1: {
+          max: 1,
+          duration: 1000,
+        },
+        group2: {
+          max: 1,
+          duration: 2000,
+        },
+      },
+    };
+
+    const rateLimitedQueue = new Queue(queueName, { limiter: limiterConfig });
+
+    const worker = new Worker(queueName, async job => {}, {
+      limiter: limiterConfig,
+    });
+    await worker.waitUntilReady();
+
+    const completed: { [index: string]: number[] } = {};
+
+    const running = new Promise((resolve, reject) => {
+      const afterJobs = after(numGroups * numJobsPerGroup, () => {
+        try {
+          // All jobs should be completed by the time all 'group3' jobs
+          // have been completed, since they use the slowest rate limit
+          // which is the default (1 job per 3 sec)
+          const timeDiff = Date.now() - startTime;
+          expect(timeDiff).to.be.gte((numJobsPerGroup - 1) * 3000);
+          expect(timeDiff).to.be.below(numJobsPerGroup * 3000);
+
+          // Test that the time diff between jobs in their groups respect
+          // the duration that was configured (within a 1-second window)
+          for (const group in completed) {
+            let prevTime = completed[group][0];
+            let lowerBound = 0;
+            if (group == 'group1') {
+              lowerBound = 1000;
+            } else if (group == 'group2') {
+              lowerBound = 2000;
+            } else if (group == 'group3') {
+              lowerBound = 3000;
+            }
+            for (let i = 1; i < completed[group].length; i++) {
+              const diff = completed[group][i] - prevTime;
+              expect(diff).to.be.below(lowerBound + 1000 + buffer);
+              expect(diff).to.be.gte(lowerBound - buffer);
+              prevTime = completed[group][i];
+            }
+          }
+          resolve();
+        } catch (err) {
+          reject(err);
+        }
+      });
+
+      queueEvents.on('completed', ({ jobId }) => {
+        const group = last(jobId.split(':'));
+        completed[group] = completed[group] || [];
+        completed[group].push(Date.now());
+
+        afterJobs();
+      });
+
+      queueEvents.on('failed', async err => {
+        await worker.close();
+        reject(err);
+      });
+    });
+
+    for (let group = 0; group < numGroups; group++) {
+      for (let i = 0; i < numJobsPerGroup; i++) {
+        const groupId = `${group + 1}`;
+        await rateLimitedQueue.add('group-specific rate test', {
+          accountId: `group${groupId}`,
+        });
+      }
+    }
+
+    await running;
+
+    const completedCount = await rateLimitedQueue.getCompletedCount();
+    expect(completedCount).to.eq(numJobsPerGroup * numGroups);
+
     await rateLimitedQueue.close();
     await worker.close();
     await queueScheduler.close();